gnosis: move functions out of main
