+++ op-chain-ops/cmd/gnosis/main.go
@@ -7,6 +7,7 @@ import (
 	"github.com/ethereum-optimism/optimism/op-node/version"
 	opservice "github.com/ethereum-optimism/optimism/op-service"
 	"github.com/ethereum-optimism/optimism/op-service/cliapp"
+	"github.com/ethereum-optimism/optimism/op-service/gnosis"
 	"github.com/urfave/cli/v2"
 )
 
@@ -23,13 +24,13 @@ func main() {
 	app.Version = VersionWithMeta
 	app.Name = "gnosis"
 	app.Usage = "tool to interact with pre-deployed gnosis safe contracts"
-	app.Flags = cliapp.ProtectFlags(GlobalFlags)
+	app.Flags = cliapp.ProtectFlags(gnosis.GlobalFlags)
 	app.Commands = []*cli.Command{
 		{
 			Name:   "send-tx",
 			Usage:  "send tx via Gnosis Safe using calldata",
-			Flags:  cliapp.ProtectFlags(SendGnosisTxFlags),
-			Action: SendGnosisTransactionCLI,
+			Flags:  cliapp.ProtectFlags(gnosis.SendGnosisTxFlags),
+			Action: gnosis.SendGnosisTransactionCLI,
 		},
 	}
 	app.Writer = os.Stdout
