+++ op-service/gnosis/cli.go
@@ -1,4 +1,4 @@
-package main
+package gnosis
 
 import (
 	"fmt"
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/ethereum-optimism/optimism/op-service/ctxinterrupt"
-	"github.com/ethereum-optimism/optimism/op-service/gnosis"
 	oplog "github.com/ethereum-optimism/optimism/op-service/log"
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
 	"github.com/ethereum/go-ethereum/common"
@@ -61,8 +60,8 @@ func SendGnosisTransactionCLI(cliCtx *cli.Context) error {
 		return fmt.Errorf("invalid calldata: %s", calldataHex)
 	}
 
-	gnosisClient, err := gnosis.NewGnosisClient(lgr, rpcUrl, privateKeys, safeAddress,
-		gnosis.WithCustomTxMgr(func(cfg *txmgr.CLIConfig) { cfg.NumConfirmations = 1 }),
+	gnosisClient, err := NewGnosisClient(lgr, rpcUrl, privateKeys, safeAddress,
+		WithCustomTxMgr(func(cfg *txmgr.CLIConfig) { cfg.NumConfirmations = 1 }),
 	)
 	if err != nil {
 		return fmt.Errorf("failed to create gnosis client: %w", err)
